requirements.txt
transformers
guess_language-spirit
python-Levenshtein
pandas~=1.4.2
numpy~=1.22.3
BeautifulSoup4~=4.11.1
conllu~=4.5.2
requests~=2.28.1
tqdm~=4.64.0
scipy~=1.7.3
matplotlib~=3.5.1

pyenchant~=3.2.2
pypinyin # Chinese
pykakasi~=2.2.1
jamotools # Korean

polyleven~=0.8
fuzzywuzzy~=0.18.0
symspellpy~=6.7.7
seaborn~=0.12.0
rapidfuzz~=2.11.1
apsw~=3.39.4.0
fasttext~=0.9.2
setuptools~=60.9.3
datasets~=2.6.0